Initial commit of Harbormaster
Harbormaster uses real-time parking information made available by
Cleveland State University to choose optimal garage for commuters.

It depends upon PHPMailer, and is intended to be run as a cron job
when your are almost at the University, so you can choose between a
preferred and backup garage.  Should capacity be really low, it will
recommend you park elsewhere so as to not waste time driving around.

Harbormaster is licensed as Apache 2.0, and contributions are welcome.

Signed-off-by: Brian Warner <brian@bdwarner.com>