Display spots as plural when != 1, handle mangled capacity data
One minor cosmetic change, so that when there's one spot available it is
displayed as "1 spot" instead of "1 spots".

Also, the Central garage data gets weird when it is full, reporting the total
capacity of the garage as 10 spots. This results in weird negative counts.
Since negative spots don't make sense, Harbormaster now just takes the max of
available spots or zero.

Signed-off-by: Brian Warner <brian@bdwarner.com>