Fix cold/heat warnings, add caches per user type.
Cold and heat warnings were broken, mainly because I forgot to ever set them.
That is fixed now.

Also, caching is now done per user type, as set by the query string. This
will make it easier to serve a different page for a web user, as compared
to email and Android users.  There is also a debugging variable called 'me'.

Signed-off-by: Brian Warner <brian@bdwarner.com>